# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "OpenAPI definition", "version": "v0" },
  "paths": {
    "/api/v1/chamados": {
      "get": {
        "tags": ["chamado-controller"],
        "operationId": "fila",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ABERTO",
                "EM_ATENDIMENTO",
                "AGUARDANDO_USUARIO",
                "RESOLVIDO",
                "FECHADO"
              ]
            }
          },
          {
            "name": "prioridade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["BAIXA", "MEDIA", "ALTA", "CRITICA"]
            }
          },
          {
            "name": "responsavelId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "categoriaId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "setorId",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "ate",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "texto",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "semResponsavel",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false }
          },
          {
            "name": "meus",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false }
          },
          {
            "name": "slaVencendo",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": false }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 20 }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "criadoEm,desc" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaResponseChamadoResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["chamado-controller"],
        "operationId": "criar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CriarChamadoRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chamado criado",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ChamadoResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/chamados/{id}/assumir": {
      "post": {
        "tags": ["chamado-controller"],
        "operationId": "assumir",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssumirChamadoRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ChamadoResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/dev/login": {
      "post": {
        "tags": ["dev-login-controller"],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UsuarioResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/chamados/{id}": {
      "get": {
        "tags": ["chamado-controller"],
        "operationId": "detalhe",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ChamadoResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["chamado-controller"],
        "operationId": "atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarChamadoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ChamadoResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": ["usuario-controller"],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/UsuarioResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/chamados/{id}/historico": {
      "get": {
        "tags": ["chamado-controller"],
        "operationId": "historico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaResponseHistoricoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chamados/meus": {
      "get": {
        "tags": ["chamado-controller"],
        "operationId": "meus",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 0 }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 20 }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "criadoEm,desc" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaResponseChamadoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categorias": {
      "get": {
        "tags": ["categoria-controller"],
        "operationId": "listar",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategoriaResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "tags": ["auth-controller"],
        "operationId": "csrf",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/CsrfResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/config": {
      "get": {
        "tags": ["auth-controller"],
        "operationId": "config",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ConfigResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CriarChamadoRequest": {
        "type": "object",
        "properties": {
          "titulo": { "type": "string", "maxLength": 200, "minLength": 0 },
          "descricao": { "type": "string", "maxLength": 10000, "minLength": 0 },
          "categoriaId": { "type": "integer", "format": "int64" },
          "prioridadeSugerida": {
            "type": "string",
            "enum": ["BAIXA", "MEDIA", "ALTA", "CRITICA"]
          },
          "solicitanteId": { "type": "integer", "format": "int64" }
        },
        "required": ["categoriaId", "descricao", "titulo"]
      },
      "ChamadoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "numero": { "type": "string" },
          "titulo": { "type": "string" },
          "descricao": { "type": "string" },
          "solicitanteId": { "type": "integer", "format": "int64" },
          "abertoPorId": { "type": ["integer", "null"], "format": "int64" },
          "categoriaId": { "type": "integer", "format": "int64" },
          "responsavelId": { "type": ["integer", "null"], "format": "int64" },
          "prioridade": {
            "type": "string",
            "enum": ["BAIXA", "MEDIA", "ALTA", "CRITICA"]
          },
          "prioridadeSugerida": {
            "type": ["string", "null"],
            "enum": ["BAIXA", "MEDIA", "ALTA", "CRITICA"]
          },
          "status": {
            "type": "string",
            "enum": [
              "ABERTO",
              "EM_ATENDIMENTO",
              "AGUARDANDO_USUARIO",
              "RESOLVIDO",
              "FECHADO"
            ]
          },
          "criadoEm": { "type": "string", "format": "date-time" },
          "atualizadoEm": { "type": "string", "format": "date-time" },
          "resolvidoEm": { "type": ["string", "null"], "format": "date-time" },
          "solucao": { "type": ["string", "null"] },
          "prazoResolucao": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "version": { "type": "integer", "format": "int64" }
        },
        "required": [
          "abertoPorId",
          "atualizadoEm",
          "categoriaId",
          "criadoEm",
          "descricao",
          "id",
          "numero",
          "prazoResolucao",
          "prioridade",
          "prioridadeSugerida",
          "resolvidoEm",
          "responsavelId",
          "solicitanteId",
          "solucao",
          "status",
          "titulo",
          "version"
        ]
      },
      "AssumirChamadoRequest": {
        "type": "object",
        "properties": { "version": { "type": "integer", "format": "int64" } },
        "required": ["version"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 1 },
          "nome": { "type": "string", "maxLength": 180, "minLength": 0 }
        },
        "required": ["email", "nome"]
      },
      "UsuarioResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "perfil": {
            "type": "string",
            "enum": ["FUNCIONARIO", "TI_AGENTE", "TI_ADMIN"]
          }
        }
      },
      "AtualizarChamadoRequest": {
        "type": "object",
        "properties": {
          "version": { "type": "integer", "format": "int64" },
          "status": {
            "type": "string",
            "enum": [
              "ABERTO",
              "EM_ATENDIMENTO",
              "AGUARDANDO_USUARIO",
              "RESOLVIDO",
              "FECHADO"
            ]
          },
          "prioridade": {
            "type": "string",
            "enum": ["BAIXA", "MEDIA", "ALTA", "CRITICA"]
          },
          "categoriaId": { "type": "integer", "format": "int64" },
          "responsavelId": { "type": "integer", "format": "int64" },
          "removerResponsavel": { "type": "boolean" },
          "solucao": { "type": "string", "maxLength": 10000, "minLength": 0 }
        },
        "required": ["version"]
      },
      "PaginaResponseChamadoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ChamadoResponse" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" }
        },
        "required": ["content", "page", "size", "totalElements", "totalPages"]
      },
      "HistoricoResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "usuarioId": { "type": "integer", "format": "int64" },
          "campo": { "type": "string" },
          "valorAnterior": { "type": "string" },
          "valorNovo": { "type": "string" },
          "criadoEm": { "type": "string", "format": "date-time" }
        }
      },
      "PaginaResponseHistoricoResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HistoricoResponse" }
          },
          "page": { "type": "integer", "format": "int32" },
          "size": { "type": "integer", "format": "int32" },
          "totalElements": { "type": "integer", "format": "int64" },
          "totalPages": { "type": "integer", "format": "int32" }
        },
        "required": ["content", "page", "size", "totalElements", "totalPages"]
      },
      "CategoriaResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "nome": { "type": "string" }
        },
        "required": ["id", "nome"]
      },
      "CsrfResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      },
      "ConfigResponse": {
        "type": "object",
        "properties": {
          "modo": { "type": "string" },
          "urlLogin": { "type": "string" }
        }
      }
    }
  }
}